+++ src/ws_server/requestHandler.ts
@@ -1,65 +1,28 @@
 import { WebSocket } from 'ws';
 import * as types from '../interfaces';
-import { wsServer } from '.';
 
-let registeredUsers: User[] = []; 
-class User {
-    id: number;
-    name: string;
-    password: string;
-    wins: number;
-    losses: number;
+import { User, findUserByName, validatePassword, updateWinners, updateSocket, registeredUsers } from './player'
 
-    constructor(name: string, password: string) {
-      this.id = registeredUsers.length;
-      this.name = name;
-      this.password = password;
-      this.wins = 0;
-      this.losses = 0;
-    }
+export const rooms: Room[] = [];
 
-    addWin() {
-        this.wins++;
-    }
+export class Room {
+  roomID: number | string;
+  roomUsers: User[];
 
-    addLoss() {
-        this.losses++;
-    }
+  constructor(user: User) {
+    this.roomID = rooms.length;
+    this.roomUsers = [user];
   }
-
-  function findUserByName(name: string): User | undefined {
-    return registeredUsers.find(user => user.name === name);
-  }
-  
-
-  function validatePassword(name: string, password: string): boolean {
-    const user = findUserByName(name);
-    if (!user) {
-      registeredUsers.push(new User(name, password))
-      return true;
-    }
-    const isValid = user.password === password;
-    return isValid;
-  }
-
-function updateWinners() {
-    let scoreTable: unknown[] = [];
-    registeredUsers.forEach(user => {
-        let scoreEntry = {'name': user.name, 'wins': user.wins};
-        scoreTable.push(scoreEntry)       
-    });
-    let response: types.reqOutputInt = new types.Reponse('update_winners', JSON.stringify(scoreTable));
-    wsServer.broadcast(JSON.stringify(response))
 }
 
 export const requestHandler = (req: types.reqInputInt, socket: WebSocket) => {
-
+  let data = (req.data) ? JSON.parse(req.data) : '';
+  let responseData;
     switch (req.type) {
         case 'reg':
-            let data = JSON.parse(req.data);
-            let responseData;
             if (validatePassword(data.name, data.password)) {
                 let loggedInUser = findUserByName(data.name) as User;
+                updateSocket(loggedInUser.id, socket);
                 responseData = new types.RegOutputData(loggedInUser.name, loggedInUser.id);
             }
             else {
@@ -70,6 +33,12 @@ export const requestHandler = (req: types.reqInputInt, socket: WebSocket) => {
             socket.send(JSON.stringify(response))
             updateWinners()
             break;
+
+        case 'create_room':
+          const plaerToAddId: User | undefined  = registeredUsers.find(user => user.socket === socket);
+          rooms.push(new Room(plaerToAddId as User));
+          console.log(rooms)
+          break;
     
         default:
             break;
